refactor: moved budget percentage chart controls to a common component
